Migration configuration update (#76)
* Provide flyway with DB settings

* Start migration from version one

Flyway doesn't see fresh postgres database on Azure as empty (maybe because of some initial functions or views that come with it). For that reason it requires us to set baselineOnMigration to true. That, however, causes another problem, which is the default baseline version (1), which comes after our initial version - 001. This is why the baseline version is set to 000.